Separate alltoall-based mpp_global_field function
When local is sliced over its compute domain (rather than the full data
domain), the input data is no longer contiguous and prevents us from
defining sensible MPI types and using MPI-based functions.

This generally has catastrophic effects, with internal function stacks
becoming corrupted and resulting in immediately SIGSEGV failures.

The alltoall-based method is now moved to a separate function which is
only called when `local` matches the data domain, which we presume to
implty that `local` is contiguous.

While not a strict contiguity test, it is sufficient until there is more
widespread support for the F2015 IS_CONTIGUOUS function.

In context of this issue, the following changes were made:

- xonly, yonly, and full scatter have been rewritten to use
  mpp_alltoall.  All four cases are now supported.

    - Some consolidation was done to streamline these cases, but there
      is far too much redundant code and more work needs to be done.

- The original mpp_do_global_field function has been restored for the
  non-contiguous case (compute slice of local).

- A new mpp_do_global_field_a2a function now contains the alltoall-based
  method.  Associated interfaces have also been created.

- The internal self-update from `local` to `global` has been removed,
  and now relies on MPI to do the transfer.

- mpp_alltoall no longer returns for single-rank and still performs the
  MPI_Alltoallw function to perform its self-transfer from `local` to
  `global`.  This is in line with the other mpp_alltoall functions.

- Non-primitive MPI datatypes are now deallocated during mpp_exit,
  rather than after each mpp_alltoall call.

- Some parti